Move external pkg to the api sub module
With moving the external pkg to the api sub module the dependency
to github.com/openstack-k8s-operators/keystone-operator can be
removed on other operators. They would then only depend on
github.com/openstack-k8s-operators/keystone-operator/api and don't
import anything used in github.com/openstack-k8s-operators/keystone-operator
which is not required.
